move macOS monitor to separate file
